Added timer which when expired will offload the dumps
Added timer which will expire for every 5 seconds and
offloads if any dumps present in the queue.

Earlier offloading one after the other but that is error
prone. As delete all dump request will log multiple errors
as it will try to delete and offload the next dump from
queue but it will not find next dump. Next dump is not
found as the queue did not get the interfacesRemoved
callback for other dumps yet as the caller thread is
blocked deleting and offloading dumps.

Added hoststate watch to avoid D-Bus calls to check if
host is running prior to every dump offload.

Added hmc state watch to avoid D-Bus calls to check if
host is running prior to every dump offload.

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: I9e844607ff8a1e4d07b845d1143086e22392a681