Check if dump is already generated before adding property watch
1) Elog, Core, Checkstop dump gets completed pretty fast
and probably done in the caller thread so progress will be set
when the dump entry is created. Adding propertywatch on the entry
is not helpful as it will never get the property changed
signal.

2) When new dump entry signal is received check if progress
is completed and offload the dump. If progress is not completed
probably it is ongoing to add property watch.

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: Ib6c6d68e8453b7d690e60991c358c68cff864395