Reattempt dump offload if PLDM is down and no callback is received
PHYP can process one dump offload at a time if there are multiple
dumps to be offloaded when system reaches to OSRunning state.

Offload handler uses offload in progress flag to ensure
one dump at a time is offloaded. This flag is reset after dump
delete request is received from PLDM.

In this case PLDM accepted dump offload request but went down
and due to checkstop. Due to this PLDM will not offload and
does not send the dump delete request.

As the flag is not reset any future offloads will not happen.

Now modifying to reset the flag when host goes down and this
way it will reattempt dump offload when host is in running state.

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: Ic0aa6209a491fe1c81a2c20a04201f0e5afcf539